// ui-packages/ui-components/lib/sidenav/sidenav.type.ts
export enum SidenavItemId {
  Files = 'Files',
  Examples = 'Examples',
  Libraries = 'Libraries',
  Reference = 'Reference',
  GenAI = 'GenAI',
  Settings = 'Settings',
}

export type SidenavItemRecord = Record<SidenavItemId, SidenavItem>;
// ...
export interface ContentLogicMap {
  [SidenavItemId.Files]: UseFilesLogic;
  [SidenavItemId.Examples]: UseExamplesLogic;
  [SidenavItemId.Libraries]: UseLibrariesLogic;
  [SidenavItemId.Reference]: UseReferenceLogic;
  [SidenavItemId.GenAI]: UseGenAILogic;
  [SidenavItemId.Settings]: UseSettingsLogic;
}

export interface HeaderLogicMap {
  [SidenavItemId.Files]: () => null;
  [SidenavItemId.Examples]: () => null;
  [SidenavItemId.Libraries]: () => {
    onLibraryUpload: () => void;
    isUploadingLibrary: boolean;
  };
  [SidenavItemId.GenAI]: () => null;
  [SidenavItemId.Reference]: () => null;
  [SidenavItemId.Settings]: () => null;
}

export interface SidenavItem {
  label: MessageDescriptor;
  labelDetails?: MessageDescriptor;
  Icon: React.FC;
  active?: boolean;
  position?: 'top' | 'bottom';
}

export type SidenavItemWithId = SidenavItem & { id: SidenavItemId };
// ...
export type UseLibrariesLogic = () => {
  onClickInclude: OnClickInclude;
  pinnedLibraries?: Libs_CreateApi;
  getCustomLibraries: GetCustomLibraries;
  getFavoriteLibraries: GetFavoriteLibraries;
  getBoards: GetBoards;
  selectedBoard?: string;
  selectedArchitecture?: string;
  canModifySketchMetadata?: boolean;
  libraryMenuHandlers: LibraryMenuHandlerDictionary;
  enableGetCustomLibraries: boolean;
  initialSelectedTab?: SidenavLibrariesIds;
  hydrateExamplesByPaths: (paths: string[]) => Promise<void>;
};
// ...
export type SidenavSections = {
  [Property in SidenavItemId]: (
    logic: ContentLogicMap[Property],
    key?: string,
  ) => JSX.Element;
};

export type HeaderMap = {
  [Property in SidenavItemId]: (
    logic: HeaderLogicMap[Property],
    key?: string,
  ) => React.ReactNode;
};
// ...
export type UseFilesLogic = () => {
  files: SelectableFileData[];
  isLoading: boolean;
  selectedFileId?: string;
  unsavedFileIds: Set<string> | undefined;
  newFileAction: (key: Key | null) => void;
  selectedFileChange: (fileId: string) => void;
  renameFile: RenameFileHandler;
  deleteFile: DeleteFileHandler;
  validateFileName: (
    newName: string,
    prevName: string,
    ext: string,
  ) => FileNameValidationResult;
  replaceFileNameInvalidCharacters: (fileName: string) => string;
  isReadOnly: boolean;
};

export type GetExamples = (enableGetExamples?: boolean) => {
  examples: Example[];
  isLoading: boolean;
};

export type UseExamplesLogic = () => {
  getExamples: GetExamples;
  initialSelectedTab?: SidenavExamplesIds;
  selectedBoard?: string;
  selectedArchitecture?: string;
  hydrateByPaths: (paths: string[]) => Promise<void>;
};
// ...
export type UseReferenceLogic = () => {
  getReferenceCategories: GetReferenceCategories;
  getReferenceItem: GetReferenceItem;
  searchReferenceItem: SearchReferenceItem;
  selectedTab: SidenavReferenceIds;
  setSelectedTab: (tab: SidenavReferenceIds) => void;
};
// ...
export type UseGenAILogic = () => {
  handleGenAiApplyCode: OnClickApplySketch;
  handleGenAiApplyFixToCode: OnClickApplyFixToSketch;
  handleApplyPatchAvailability: ApplyPatchAvailability;
  scrollToLine: (line: number, fileName?: string) => void;
  errorLines?: number[];
  genAiMessageUsageExceeded?: boolean;
  shouldDisplayAiLimitBanner?: boolean;
  aiMessagesRemaining?: number;
  linksEnabled: boolean;
  openGenAIPolicyTermsDialog: () => void;
  onUserInput: (text: string) => void;
  maxContentLength: number;
};

export type UseSettingsLogic = () => {
  getPreference: (id: Preferences) => PreferenceValue | undefined;
  setPreference: (id: Preferences, value: PreferenceValue) => void;
  restorePreferences: () => void;
  themeContext: {
    theme: string;
    setTheme: React.Dispatch<React.SetStateAction<string>>;
    isDarkModeOs: boolean;
  };
  saveAllFiles: () => void;
  handleOptOut?: () => void;
};